Make numeric predicates work with rationals
Generalise numeric predicates so that they work with any of Nat, TInt or
Rat.  However this has required restricting the type being refined
slightly further from (Num, Ord) to Real.  This is in order to convert
to Rational for the comparison.

Weakness: error messages are displayed as double because the
prettyprinter library does not have a Pretty instance for Ratio a.  We
could include an orphan instance but again this should probably instead
be changed upstream.

This would resolve #10 when typenums hits hackage.